+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#137](https://github.com/zpix1/yt-anti-translate/issues/137) Videos description in searches sometimes shows older videos description
 - Add missing selector to handle branding of feutured channels and sidebar channels
 - Fix video description not translating of fundraising videos if viewport is below 1000x500
+- Added missing selector to handle description of dismissable featured channel video
 
 ## [1.19.13] - 2025-09-20