+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.20.0] - 2025-10-xx
+## [1.20.0] - 2025-10-13
 
 ### New Feature